+++ packages/assets/src/index.ts
@@ -1,4 +1,5 @@
 import { CDN_BASE_URL as EMAIL_CDN_BASE_URL, EMAIL_MANIFEST } from '@porsche-design-system/email';
+import { CDN_BASE_URL as FALLBACKS_CDN_BASE_URL, FALLBACKS_MANIFEST } from '@porsche-design-system/fallbacks';
 import { CDN_BASE_URL as FONTS_CDN_BASE_URL, FONTS_MANIFEST } from '@porsche-design-system/fonts';
 import { CDN_BASE_URL as ICONS_CDN_BASE_URL, ICONS_MANIFEST, ICON_NAMES } from '@porsche-design-system/icons';
 import { CDN_BASE_URL as MARQUES_CDN_BASE_URL, MARQUES_MANIFEST } from '@porsche-design-system/marque';
@@ -9,6 +10,8 @@ import { FONT_FACE_CDN_URL } from '@porsche-design-system/styles';
 export {
   EMAIL_CDN_BASE_URL,
   EMAIL_MANIFEST,
+  FALLBACKS_CDN_BASE_URL,
+  FALLBACKS_MANIFEST,
   FONTS_CDN_BASE_URL,
   FONTS_MANIFEST,
   ICONS_CDN_BASE_URL,